fix: refactor Helm templateArgs to handle overrides and file creation
